fix: Race condition in nse/ns Find(watch=true)
Related issue: gh-573
The fault was introduced by commit he commit bb0a6d4 (fix for GH-533).
To fix the race condition, now wait until the channel is added to subscribers before returning.

Other problem introduced is the blocking sendEvent(). This combined with the serialized executor creates a deadlock risk. If one subscriber is slow, it blocks the entire executor and prevents other events from being processed. The select was readded to sendEvent().

Signed-off-by: Laszlo Kiraly <laszlo.kiraly@est.tech>